refactor: extract `initApp` from `build()`
