+++ packages/nuxt3/src/core/nuxt.ts
@@ -8,6 +8,7 @@ import componentsModule from '../components/module'
 import autoImportsModule from '../auto-imports/module'
 import { distDir, pkgDir } from '../dirs'
 import { version } from '../../package.json'
+import { initApp } from './app'
 import { initNitro } from './nitro'
 import { addModuleTranspiles } from './modules'
 
@@ -73,6 +74,8 @@ async function initNuxt (nuxt: Nuxt) {
   await nuxt.callHook('modules:done', { nuxt } as ModuleContainer)
 
   await addModuleTranspiles()
+
+  await initApp(nuxt)
 
   await nuxt.callHook('ready', nuxt)
 }
